+++ src/types/services.ts
@@ -60,20 +60,3 @@ export interface Hotel {
 export interface HotelFormType {
   name: string;
 }
-
-export interface AgencyUser{
-  id: number
-  name: string
-  email: string
-  password: string
-  role: string
-  agencyId: number
-}
-
-export interface AgencyUserFormType {
-  name: string
-  email: string
-  password: string
-  role: string
-  agencyId: number
-}
